BaseModel: Interpret literal `true` values as true
When invoked with the literal value `true`, the built-in PHP function
`strtolower` returns the string value `"1"`. Previously, the
sanitization logic for model attributes did not account for this, so
such attributes were sanitized to the database's equivalent encoding for
"false" values.

When sanitizing model attributes, ensure that boolean attributes
specified with the literal PHP value `true` are persisted with the
database's equivalent encoding for "true" values.